backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
Flask-CORS
PyJWT
bcrypt
python-dotenv
